+++ bin/VisualizzaEmozioni.fxml
@@ -10,10 +10,10 @@
    <children>
   <AnchorPane layoutY="-4.0" minHeight="-Infinity" minWidth="-Infinity" prefHeight="550.0" prefWidth="750.0" AnchorPane.bottomAnchor="-4.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="-4.0">
          <children>
-            <Rectangle arcHeight="20.0" arcWidth="20.0" fill="#ffffff00" height="350.0" layoutX="17.0" layoutY="148.0" stroke="WHITE" strokeLineJoin="ROUND" strokeType="INSIDE" width="716.0" />
-            <TableView fx:id="tabellaEmozioni" layoutX="20.0" layoutY="163.0" prefHeight="320.0" prefWidth="710.0" />
+            <Rectangle arcHeight="20.0" arcWidth="20.0" fill="#ffffff00" height="350.0" layoutX="17.0" layoutY="148.0" stroke="WHITE" strokeLineJoin="ROUND" strokeType="INSIDE" width="710.0" />
+            <TableView fx:id="tabellaEmozioni" layoutX="17.0" layoutY="163.0" prefHeight="320.0" prefWidth="710.0" />
             <Button layoutX="297.0" layoutY="510.0" mnemonicParsing="false" onAction="#tornaAlPrecedente" prefHeight="26.0" prefWidth="156.0" text="torna alla ricerca" />
-            <Rectangle arcHeight="20.0" arcWidth="20.0" fill="#ffffff00" height="87.0" layoutX="17.0" layoutY="38.0" stroke="WHITE" strokeLineJoin="ROUND" strokeType="INSIDE" width="716.0" />
+            <Rectangle arcHeight="20.0" arcWidth="20.0" fill="#ffffff00" height="87.0" layoutX="17.0" layoutY="38.0" stroke="WHITE" strokeLineJoin="ROUND" strokeType="INSIDE" width="710.0" />
             <TextArea fx:id="infoCanzone" editable="false" layoutX="22.0" layoutY="47.0" prefHeight="68.0" prefWidth="700.0" />
          </children></AnchorPane>
    </children>
