+++ data/Emozioni.dati.csv
@@ -4,4 +4,3 @@ s,,1,,4,,pippo,,5,, ,,0,, ,,3,, ,,4,, ,,4,,pippo,,4,, ,,4,, ,,4,,arr,,
 o,,1,,4,,rrrrrrrrrr,,0,, ,,0,, ,,0,, ,,4,, ,,3,, ,,3,, ,,0,, ,,0,, 
 s,,272,,3,, ,,0,, ,,0,, ,,0,, ,,0,, ,,0,, ,,0,, ,,0,, ,,0,, 
 a,,215,,4,, ,,2,, ,,0,, ,,4,, ,,0,, ,,3,, ,,1,, ,,1,, ,,1,, 
-emotional listener,,1,,0,, ,,2,,prova,,3,, ,,0,, ,,2,,prova,,3,,prova,,3,, ,,3,,prova,,0,,
