+++ data/UtentiRegistrati.dati.csv
@@ -3,4 +3,3 @@ a,a,a,a,a,a,a
 m,m,m,m,m,m,m
 o,o,o,o,o,o,o
 s,s,s,s,s,s,s
-prova,prova,prova,prova;prova;prova;prova;prova;,prova,emotional listener,emotionalsongs
