--- src/emotionalsongs/.gitignore
@@ -1,1 +1,0 @@
-/Main.java
