+++ src/emotionalsongs/RicercaInRepositoryController.java
@@ -7,9 +7,14 @@ package emotionalsongs;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
+
 import javafx.collections.FXCollections;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
 
 /**
  * Classe che si occupa della gestione della scena descritta da RicercaInRepository.fxml
