+++ src/emotionalsongs/SelezionaPlaylistController.java
@@ -6,7 +6,10 @@ package emotionalsongs;
 
 import java.io.IOException;
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.Random;
 import java.util.ResourceBundle;
+
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
